Correct explanation of real expansion
I thought I submitted a pull request earlier today, but then it disappeared, so I re-did the changes and am submitting a new one.

The reason why the factor \sqrt{2} appears is not for "convenience" - it is an orthogonality requirement when we switch to an entirely real basis (instead of the real+imaginary spherical harmonics basis). This is shown in Appendix A of the Denovo manual.